+++ src/isNumber/tests.js
@@ -1,7 +1,0 @@
-import isNumber from './index';
-import test from 'ava';
-
-test('Test isNumber()', t => {
-  t.truthy(isNumber(1), 'Results returned true');
-  t.falsy(isNumber('1'), 'Results returned false');
-});