+++ src/mean/tests.js
@@ -7,10 +7,3 @@ test('Test base functionality', t => {
   t.truthy(results);
   t.is(results, 2);
 });
-
-test('Test array functionality', t => {
-  const results = mean([1, 2, 3, 2]);
-
-  t.truthy(results);
-  t.is(results, 2);
-});